pyproject.toml
[tool.poetry]
name = "torappu"
version = "0.1.0"
description = ""
authors = ["daflyinbed <1677759063@qq.com>"]
readme = "README.md"

packages = [
    { include = "torappu/core" }
]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.24.1"
sqlalchemy = "^2.0.15"
unitypy = "^1.10.1"
flatbuffers = "^23.5.26"
pycryptodome = "^3.18.0"
bson = "^0.5.10"
apscheduler = "^3.10.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pycln = "^2.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
src_paths = ["nonebot", "tests"]
extra_standard_library = ["typing_extensions"]

[tool.pycln]
path = "."
all = false